+++ backend/filtering/repository_test.go
@@ -1,0 +1,32 @@
+package filtering
+
+import (
+	"context"
+	"os"
+	"testing"
+
+	e "github.com/pkg/errors"
+	log "github.com/sirupsen/logrus"
+	_ "gorm.io/driver/postgres"
+	"gorm.io/gorm"
+
+	"github.com/highlight-run/highlight/backend/util"
+)
+
+var DB *gorm.DB
+
+// Gets run once; M.run() calls the tests in this file.
+func TestMain(m *testing.M) {
+	dbName := "highlight_testing_db"
+	testLogger := log.WithContext(context.TODO()).WithFields(log.Fields{"DB_HOST": os.Getenv("PSQL_HOST"), "DB_NAME": dbName})
+	var err error
+	DB, err = util.CreateAndMigrateTestDB(dbName)
+	if err != nil {
+		testLogger.Error(e.Wrap(err, "error creating testdb"))
+	}
+	code := m.Run()
+	os.Exit(code)
+}
+
+func TestUpdateProjectFilters(t *testing.T) {
+}
